Performed migration from endpoints to Minimal API
