+++ EventSourcing.NetCore.sln
@@ -325,6 +325,14 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "14-Projections.SingleStream
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EventPipelines.EventStoreDB", "Sample\EventPipelines\EventPipelines.EventStoreDB\EventPipelines.EventStoreDB.csproj", "{9B7FBEE7-99AE-4A94-8E36-FBE1A62BEEE0}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Warehouse.MinimalAPI", "Warehouse.MinimalAPI", "{005FC7CE-97BA-47F7-982E-63C46327F78C}"
+	ProjectSection(SolutionItems) = preProject
+		Sample\Warehouse.MinimalAPI\docker-compose.yml = Sample\Warehouse.MinimalAPI\docker-compose.yml
+		Sample\Warehouse.MinimalAPI\README.md = Sample\Warehouse.MinimalAPI\README.md
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Warehouse.Api", "Sample\Warehouse.MinimalAPI\Warehouse.Api\Warehouse.Api.csproj", "{723AD8CE-E3EF-40F6-9576-CCF2C3DB504E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -743,6 +751,10 @@ Global
 		{9B7FBEE7-99AE-4A94-8E36-FBE1A62BEEE0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9B7FBEE7-99AE-4A94-8E36-FBE1A62BEEE0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9B7FBEE7-99AE-4A94-8E36-FBE1A62BEEE0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{723AD8CE-E3EF-40F6-9576-CCF2C3DB504E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{723AD8CE-E3EF-40F6-9576-CCF2C3DB504E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{723AD8CE-E3EF-40F6-9576-CCF2C3DB504E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{723AD8CE-E3EF-40F6-9576-CCF2C3DB504E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -874,6 +886,8 @@ Global
 		{7A990FD1-4935-4F85-912C-D4B870232856} = {65F6E2BE-B2D4-4E56-B0CB-3062C4882B9E}
 		{A5A2A4CB-08F6-49AE-B044-DD9C5F8B0F48} = {65F6E2BE-B2D4-4E56-B0CB-3062C4882B9E}
 		{9B7FBEE7-99AE-4A94-8E36-FBE1A62BEEE0} = {D9799DB3-9D11-4909-8133-64CD96F6E1AC}
+		{005FC7CE-97BA-47F7-982E-63C46327F78C} = {A7186B6B-D56D-4AEF-B6B7-FAA827764C34}
+		{723AD8CE-E3EF-40F6-9576-CCF2C3DB504E} = {005FC7CE-97BA-47F7-982E-63C46327F78C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {A5F55604-2FF3-43B7-B657-4F18E6E95D3B}
