+++ Sample/Warehouse.MinimalAPI/Warehouse.Api/Products/Product.cs
@@ -1,0 +1,57 @@
+﻿using System;
+using System.Text.Json.Serialization;
+using System.Text.RegularExpressions;
+
+namespace Warehouse.Api.Products;
+
+internal class Product
+{
+    public Guid Id { get; set; }
+
+    /// <summary>
+    /// The Stock Keeping Unit (SKU), i.e. a merchant-specific identifier for a product or service, or the product to which the offer refers.
+    /// </summary>
+    /// <returns></returns>
+    public SKU Sku { get; set; } = default!;
+
+    /// <summary>
+    /// Product Name
+    /// </summary>
+    public string Name { get; set; } = default!;
+
+    /// <summary>
+    /// Optional Product description
+    /// </summary>
+    public string? Description { get; set; }
+
+    private Product(){}
+
+    public Product(Guid id, SKU sku, string name, string? description)
+    {
+        Id = id;
+        Sku = sku;
+        Name = name;
+        Description = description;
+    }
+}
+
+public record SKU
+{
+    public string Value { get; init; }
+
+    [JsonConstructor]
+    public SKU(string value)
+    {
+        Value = value;
+    }
+
+    public static SKU Create(string? value)
+    {
+        if (value == null)
+            throw new ArgumentNullException(nameof(SKU));
+        if (string.IsNullOrWhiteSpace(value) || !Regex.IsMatch(value, "[A-Z]{2,4}[0-9]{4,18}"))
+            throw new ArgumentOutOfRangeException(nameof(SKU));
+
+        return new SKU(value);
+    }
+}
