--- Sample/Warehouse.MinimalAPI/Warehouse/Products/Primitives/SKU.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Text.Json.Serialization;
-using System.Text.RegularExpressions;
-
-namespace Warehouse.Products.Primitives;
-
-public record SKU
-{
-    public string Value { get; init; }
-
-    [JsonConstructor]
-    private SKU(string value)
-    {
-        Value = value;
-    }
-
-    public static SKU Create(string? value)
-    {
-        if (value == null)
-            throw new ArgumentNullException(nameof(SKU));
-        if (string.IsNullOrWhiteSpace(value) || !Regex.IsMatch(value, "[A-Z]{2,4}[0-9]{4,18}"))
-            throw new ArgumentOutOfRangeException(nameof(SKU));
-
-        return new SKU(value);
-    }
-}
