+++ Sample/Warehouse.MinimalAPI/docker-compose.yml
@@ -1,0 +1,33 @@
+version: "3"
+services:
+    #######################################################
+    #  Postgres
+    #######################################################
+    postgres:
+        image: clkao/postgres-plv8
+        container_name: postgres
+        environment:
+            POSTGRES_PASSWORD: Password12!
+        ports:
+            - "5432:5432"
+        networks:
+            - pg_network
+
+    pgadmin:
+        image: dpage/pgadmin4
+        container_name: pgadmin
+        environment:
+            PGADMIN_DEFAULT_EMAIL: ${PGADMIN_DEFAULT_EMAIL:-admin@pgadmin.org}
+            PGADMIN_DEFAULT_PASSWORD: ${PGADMIN_DEFAULT_PASSWORD:-admin}
+        ports:
+            - "${PGADMIN_PORT:-5050}:80"
+        networks:
+            - pg_network
+
+networks:
+    pg_network:
+        driver: bridge
+
+volumes:
+    postgres:
+    pgadmin:
